+++ src/backend/resolvers/mutations/user/delete.js
@@ -1,7 +1,7 @@
 const deleteUser = async (
   parent,
   { input: { _id } },
-  { mongoClient, coreRoles, coreUsers, pulseDevDb, auth0 },
+  { mongoClient, coreRoles, coreUsers, pulseDevDb, pulseProdDb, auth0 },
   info
 ) => {
   // ! auth0
@@ -35,12 +35,14 @@ const deleteUser = async (
 
     // Delete user from pulse-dev users collection
     await pulseDevDb.collection('users').deleteOne({ _id }, { session })
+    await pulseProdDb.collection('users').deleteOne({ _id }, { session })
+
     console.log('user removed from pulse-dev users')
+    console.log('user removed from pulse-prod users')
 
     console.log('user no longer has sitemap or resources access')
 
     // delete user from source collection
-
     await coreUsers.findOneAndDelete({ _id }, { session })
     console.log(`${deletedUser.username} successfully deleted\n`)
   })